pyproject.toml
[tool.poetry]
name = "mqttc"
version = "0.1.0"
description = ""
authors = ["raushanraja <vraushanraja@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
paho-mqtt = "^1.6.1"
pytelegrambotapi = "^4.14.0"
python-dotenv = "^1.0.0"
aiohttp = "^3.8.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
